requirements.txt
langdetect==1.0.9
pandas>=1.1.0
sentence-transformers==2.0.0
Django>=3.2.6
django-environ>=0.4.5
psycopg2-binary>=2.9.1
google-cloud-secret-manager>=2.7.0
datefinder>=0.7.1
deep-translator>=1.5.0
django-cors-headers==3.7.0
django-cors-middleware==1.5.0
django-crispy-forms==1.12.0
django-filter==2.4.0
django-oauth-toolkit==1.5.0
djangorestframework==3.12.4
pathlib==1.0.1
gunicorn==20.1.0
